LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/iframe.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/post-to-owner.py
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/post-to-owner.py
This is a link!This is a link!This is a link!

Harness Error (TIMEOUT), message = null

PASS localhost -> localhost:9443 iframe: forced 
TIMEOUT localhost -> www.localhost:9443 iframe: forced Test timed out
TIMEOUT localhost -> www.127.0.0.1:9443 iframe: forced Test timed out
PASS localhost -> localhost:9443 iframe: user-activated 
TIMEOUT localhost -> www.localhost:9443 iframe: user-activated Test timed out
TIMEOUT localhost -> www.127.0.0.1:9443 iframe: user-activated Test timed out
FAIL localhost -> localhost:9443 iframe: forced: sec-fetch-dest assert_equals: expected "iframe" but got ""
FAIL localhost -> localhost:9443 iframe: forced: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL localhost -> localhost:9443 iframe: forced: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS localhost -> localhost:9443 iframe: forced: sec-fetch-user 
FAIL localhost -> localhost:9443 iframe: user-activated: sec-fetch-dest assert_equals: expected "iframe" but got ""
FAIL localhost -> localhost:9443 iframe: user-activated: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL localhost -> localhost:9443 iframe: user-activated: sec-fetch-site assert_equals: expected "same-origin" but got ""
FAIL localhost -> localhost:9443 iframe: user-activated: sec-fetch-user assert_equals: expected "?1" but got ""
